requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
gunicorn==21.2.0
Flask-SQLAlchemy==3.0.5
python-dotenv==1.0.0 